\name{STTm}
\alias{STTm}
\docType{data}
\title{
Soft tissue tumour dataset 
}
\description{
STTm is a matrix containing a subset of the soft tissue tumour gene expression study dataset published by Nielsen et al. (2002). The matrix constructed from 31 columns and 5520 rows, samples and genes respectively. Samples came from three different tumour types: gastrointestinal stromal tumor (GIST), leiomyosarcoma (LEIO) and synovial sarcoma (SARC).
}
\source{
The original, full dataset is downloadable from the Gene Expression Omnibus database (GSE3443) and from Stanford Microarray Database (http://smd.stanford.edu). 
}
\references{
Nielsen T. et al. (2002) Molecular characterisation of soft tissue tumours: A gene expression study. Lancet 359:1301-1307.

Mahoney M. W. and Drineas P. (2009) CUR matrix decompositions for improved data analysis. PNAS, 106(3):697-702

Andras Bodor, Istvan Csabai, Michael W Mahoney and Norbert Solymosi
rCUR:an R package for CUR matrix decomposition
BMC Bioinformatics 2012, 13:103
doi:10.1186/1471-2105-13-103
}
\examples{
data(STTm)
}
\keyword{datasets}
